# api-server/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/build/settings/{applicationId}:
        get:
            tags:
                - BuildSettingsService
            operationId: BuildSettingsService_GetBuildSettings
            parameters:
                - name: applicationId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.build.settings.BuildSettings'
        post:
            tags:
                - BuildSettingsService
            operationId: BuildSettingsService_ApplyBuildSettings
            parameters:
                - name: applicationId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.build.settings.BuildSettings'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
components:
    schemas:
...
        api.v1.build.settings.BuildSettings:
            required:
                - applicationId
                - dockerfile
            type: object
            properties:
                applicationId:
                    type: string
                dockerfile:
                    type: string
...